+++ frontend/webapp/containers/overview/index.tsx
@@ -4,4 +4,5 @@ export { SourcesContainer } from "./sources/sources";
 export { NewDestinationFlow } from "./destination/new.destination.flow";
 export { NewDestinationForm } from "./destination/new.destination.form";
 export { UpdateDestinationFlow } from "./destination/update.destination.flow";
-
+export { UpdateSourceForm } from "./sources/update.source.form";
+export { SourcesListContainer } from "./sources/sources.list.container";
