+++ frontend/webapp/containers/overview/sources/sources.styled.tsx
@@ -22,3 +22,29 @@ export const ButtonWrapper = styled.div`
   right: 32px;
   top: 40px;
 `;
+
+export const ManageSourcePageContainer = styled.div`
+  padding: 32px;
+`;
+
+export const BackButtonWrapper = styled.div`
+  display: flex;
+  width: fit-content;
+  align-items: center;
+  cursor: pointer;
+  p {
+    cursor: pointer !important;
+  }
+`;
+
+export const FieldWrapper = styled.div`
+  height: 36px;
+  width: 348px;
+  margin-bottom: 64px;
+`;
+
+export const SaveSourceButtonWrapper = styled.div`
+  margin-top: 48px;
+  height: 36px;
+  width: 362px;
+`;
